refactor(BestScore): 현 랭킹 서비스를 레거시와 분리한다
