+++ src/main/java/com/snackgame/server/rank/applegame/BestScoreRankingService.java
@@ -8,16 +8,11 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import com.snackgame.server.applegame.domain.game.AppleGame;
-import com.snackgame.server.applegame.domain.game.AppleGames;
 import com.snackgame.server.applegame.event.GameEndEvent;
-import com.snackgame.server.applegame.exception.NoRankingYetException;
 import com.snackgame.server.member.domain.AccountType;
 import com.snackgame.server.member.domain.Member;
 import com.snackgame.server.member.domain.MemberRepository;
 import com.snackgame.server.rank.applegame.controller.dto.RankResponseV2;
-import com.snackgame.server.rank.applegame.controller.dto.RankingResponse;
-import com.snackgame.server.rank.applegame.dao.SessionRankingDao;
-import com.snackgame.server.rank.applegame.dao.dto.RankingDto;
 import com.snackgame.server.rank.applegame.domain.BestScore;
 import com.snackgame.server.rank.applegame.domain.BestScores;
 import com.snackgame.server.rank.applegame.domain.Season;
@@ -31,10 +26,7 @@ public class BestScoreRankingService {
 
     private static final int RANK_PAGE_SIZE = 50;
 
-    private final SessionRankingDao sessionRankingDao;
-    private final AppleGames appleGames;
     private final BestScores bestScores;
-
     private final MemberRepository memberRepository;
     private final SeasonRepository seasonRepository;
 
@@ -79,32 +71,5 @@ public class BestScoreRankingService {
                         new BestScore(owner.getId(), season.getId())
                 ))
                 .renewWith(appleGame);
-    }
-
-    // TODO: Deprecated 된 기존 방식을 개선 정도(MTT) 테스트 후 제거
-    @Deprecated(forRemoval = true)
-    @Transactional(readOnly = true)
-    public List<RankingResponse> getEntireRankings() {
-        List<RankingDto> top50Rankings = sessionRankingDao.selectTopsByScoreIn(RANK_PAGE_SIZE);
-
-        return top50Rankings.stream()
-                .map(this::getResponseOf)
-                .collect(Collectors.toList());
-    }
-
-    @Deprecated(forRemoval = true)
-    @Transactional(readOnly = true)
-    public RankingResponse getBestRankingOf(Long memberId) {
-        return sessionRankingDao.selectBestByScoreOf(memberId)
-                .map(this::getResponseOf)
-                .orElseThrow(NoRankingYetException::new);
-    }
-
-    @Deprecated(forRemoval = true)
-    private RankingResponse getResponseOf(RankingDto rankingDto) {
-        return RankingResponse.of(
-                rankingDto.getRanking(),
-                appleGames.getBy(rankingDto.getSessionId())
-        );
     }
 }
